requirements.txt
colorama==0.4.6
google_api_python_client==2.95.0
google_auth_oauthlib==1.0.0
passlib==1.7.4
protobuf==4.24.3
rich==13.6.0
setuptools==65.5.0
